script: Move srcset-related structs and functions out of HTMLImageElement
This is a prerequisite for implementing the `imagesrcset` attribute on `<link>` elements. Not 100% on the final placement of this file or the visibility changes (I just did the simplest changes to get it compiling), some feedback would be appreciated.
