+++ components/script/test.rs
@@ -59,12 +59,6 @@ pub mod size_of {
     }
 }
 
-pub mod srcset {
-    pub use crate::dom::html::htmlimageelement::{
-        Descriptor, ImageSource, parse_a_srcset_attribute,
-    };
-}
-
 pub mod timeranges {
     pub use crate::dom::timeranges::TimeRangesContainer;
 }
